+++ src/lib.rs
@@ -1,3 +1,4 @@
 mod network_adapter;
+mod encoding;
 pub mod network;
 pub mod events;